login system

+++ src/graphql/user/resolvers.js
@@ -1,3 +1,6 @@
+import { AuthenticationError } from 'apollo-server';
+import { checkOwner } from '../login/utils/auth-functions';
+
 const user = async (_, { id }, { dataSources }) => {
   const user = dataSources.userApi.getUser(id);
   return user;
@@ -18,11 +21,19 @@ const createUser = async (_, { data }, { dataSources }) => {
   return dataSources.userApi.createUser(data);
 };
 
-const updateUser = async (_, { userId, data }, { dataSources }) => {
+const updateUser = async (
+  _,
+  { userId, data },
+  { dataSources, loggedUserId },
+) => {
+  checkOwner(userId, loggedUserId);
+
   return dataSources.userApi.updateUser(userId, data);
 };
 
-const deleteUser = async (_, { userId }, { dataSources }) => {
+const deleteUser = async (_, { userId }, { dataSources, loggedUserId }) => {
+  checkOwner(userId, loggedUserId);
+
   return dataSources.userApi.deleteUser(userId);
 };
